+++ src/clj/spring_lobby.clj
@@ -723,7 +723,9 @@
                          (do
                            (log/info "Adding task to search springfiles for map" battle-map)
                            {::task-type ::search-springfiles
-                            :springname battle-map})
+                            :springname battle-map
+                            :resource-type ::map
+                            :spring-isolation-dir spring-root})
                          (and battle-map
                               (not map-importable)
                               (not map-downloadable)
@@ -2561,7 +2563,59 @@
 
 
 (defmethod event-handler ::select-battle [{:fx/keys [event] :keys [server-key]}]
-  (swap! *state assoc-in [:by-server server-key :selected-battle] (:battle-id event)))
+  (let [battle-id (:battle-id event)
+        state (swap! *state update-in [:by-server server-key]
+                (fn [{:keys [users] :as server-data}]
+                  (let [battle-id (:battle-id event)
+                        {:keys [host-username]} (get-in server-data [:battles battle-id])]
+                    (cond-> (assoc server-data :selected-battle battle-id)
+                            (get-in users [host-username :client-status :bot])
+                            (assoc-in [:channels (u/user-channel-name host-username) :capture-until] (+ (u/curr-millis) 500))))))
+        {:keys [client-data users] :as server-data} (get-in state [:by-server server-key])
+        {:keys [host-username]} (get-in server-data [:battles battle-id])
+        is-bot (get-in users [host-username :client-status :bot])
+        channel-name (u/user-channel-name host-username)]
+    (future
+      (when is-bot
+        (log/info "Capturing chat from" channel-name)
+        @(event-handler
+           {:event/type ::send-message
+            :channel-name channel-name
+            :client-data client-data
+            :message (str "!status battle")})
+        (async/<!! (async/timeout 500))
+        (let [path [:by-server server-key :channels channel-name :capture]
+              [old-state _new-state] (swap-vals! *state assoc-in path nil)
+              captured (get-in old-state path)
+              parsed
+              (if (string/starts-with? captured "\nBattle lobby is empty")
+                []
+                (->> (string/split captured #"\n")
+                     rest
+                     rest
+                     rest
+                     (take-while #(not (string/starts-with? % "=====================")))
+                     (map #(string/split % #"\s+(?![\?\(])"))
+                     (remove #(< (count %) 7))
+                     (map (fn [vs]
+                            (if (= 8 (count vs))
+                              vs
+                              (concat (take 3 vs) [""] (drop 3 vs)))))
+                     (map (partial zipmap [:username :ally :id :clan :ready :rank :skill :user-id]))
+                     (map
+                       (fn [{:keys [ally id] :as u}]
+                         (let [id (some-> id u/to-number int dec)
+                               ally (some-> ally u/to-number int dec)]
+                           (assoc u :battle-status {:ally ally :id id}))))))]
+          (log/info "Captured chat from" channel-name ":" captured)
+          (swap! *state update-in [:by-server server-key :battles battle-id :users]
+            (fn [users]
+              (reduce
+                (fn [m u]
+                  (update m (:username u) merge u))
+                users
+                parsed))))))))
+
 
 (defmethod event-handler ::select-scenario [{:fx/keys [event]}]
   (swap! *state assoc :selected-scenario (:scenario event)))
@@ -3773,7 +3827,15 @@
                    (apply-battle-status-changes client-data (assoc id :is-me is-me) opts status-changes))))
              doall))
       (catch Exception e
-        (log/error e "Error updating to" n "teams")))))
+        (log/error e "Error updating to" n "teams")))
+    (when-not client-data
+      (let [{:keys [battle users username]} (get-in @*state [:by-server :local])]
+        (event-handler
+          (assoc e
+                 :event/type ::battle-fix-colors
+                 :battle battle
+                 :users users
+                 :username username))))))
 
 (defmethod event-handler ::battle-teams-ffa
   [e]
@@ -3952,19 +4014,20 @@
 (defmethod event-handler ::modoption-change
   [{:keys [am-host client-data modoption-key modoption-type option-key singleplayer] :fx/keys [event] :as e}]
   (let [value (modoption-value modoption-type event)
-        option-key (or option-key "modoptions")]
+        option-key (or option-key "modoptions")
+        modoption-key-str (name modoption-key)]
     (if singleplayer
       (swap! *state
              (fn [state]
                (-> state
-                   (assoc-in [:by-server :local :scripttags "game" option-key modoption-key] (str event))
-                   (assoc-in [:by-server :local :battle :scripttags "game" option-key modoption-key] (str event)))))
+                   (assoc-in [:by-server :local :scripttags "game" option-key modoption-key-str] (str event))
+                   (assoc-in [:by-server :local :battle :scripttags "game" option-key modoption-key-str] (str event)))))
       (if am-host
-        (message/send-message *state client-data (str "SETSCRIPTTAGS game/" option-key "/" (name modoption-key) "=" value))
+        (message/send-message *state client-data (str "SETSCRIPTTAGS game/" option-key "/" modoption-key-str "=" value))
         (event-handler
           (assoc e
                  :event/type ::send-message
-                 :message (str "!bSet " (name modoption-key) " " value)))))))
+                 :message (str "!bSet " modoption-key-str " " value)))))))
 
 (defmethod event-handler ::show-ai-options-window
   [{:keys [bot-name bot-username bot-version server-key]}]
@@ -4417,11 +4480,16 @@
        :mirrors mirrors})))
 
 (defmethod task-handler ::search-springfiles
-  [{:keys [springname] :as e}]
+  [{:keys [download-if-found springname] :or {download-if-found true} :as e}]
   (if-not (string/blank? springname)
     (let [search-result (search-springfiles e)]
       (log/info "Found details for" springname "on springfiles" search-result)
       (swap! *state assoc-in [:springfiles-search-results springname] search-result)
+      (when download-if-found
+        (task/add-task! *state
+          (assoc e
+                 ::task-type ::download-springfiles
+                 :search-result search-result)))
       search-result)
     (log/warn "No springname to search springfiles" e)))